refactor: reorganize project file structure
