+++ test/widget_test.dart
@@ -8,12 +8,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:todo/services/storage_service.dart';
 import 'package:todo/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
+
+    final storageService = await StorageService.init();
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(MyApp(storageService: storageService));
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
